+++ src/OpenTelemetry.Instrumentation.SqlClient/Implementation/SqlClientDiagnosticListener.cs
@@ -80,6 +80,24 @@ namespace OpenTelemetry.Instrumentation.SqlClient.Implementation
 
                         if (activity.IsAllDataRequested)
                         {
+                            try
+                            {
+                                if (this.options.Filter?.Invoke(command) == false)
+                                {
+                                    SqlClientInstrumentationEventSource.Log.CommandIsFilteredOut(activity.OperationName);
+                                    activity.IsAllDataRequested = false;
+                                    activity.ActivityTraceFlags &= ~ActivityTraceFlags.Recorded;
+                                    return;
+                                }
+                            }
+                            catch (Exception ex)
+                            {
+                                SqlClientInstrumentationEventSource.Log.CommandFilterException(ex);
+                                activity.IsAllDataRequested = false;
+                                activity.ActivityTraceFlags &= ~ActivityTraceFlags.Recorded;
+                                return;
+                            }
+
                             _ = this.connectionFetcher.TryFetch(command, out var connection);
                             _ = this.databaseFetcher.TryFetch(connection, out var database);