+++ src/OpenTelemetry.Instrumentation.SqlClient/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Unreleased
 
-* Updated to use Activity native support from `System.Diagnostics.DiagnosticSource`
+* Use `Activity.Status` and `Activity.StatusDescription` properties instead of
+  `OpenTelemetry.Trace.Status` and `OpenTelemetry.Trace.Status.Description`
+  respectively.
   to set activity status.
   ([#3118](https://github.com/open-telemetry/opentelemetry-dotnet/issues/3118))
   ([#3751](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3751))